requirements.txt
Flask==3.1.0
flask-supabase==0.2.1
python-dotenv==1.0.1